Rename the library from rspec-sad to rspec-search-and-destroy
It is reasonable to have the command be short, but it's better to have
the library code all be consistent.
