+++ lib/rspec-search-and-destroy.rb
